feat: making it work with existing links and deleting old stuff
