+++ sprout/templates/projects-id-metadata-view.html
@@ -74,7 +74,7 @@
   </nav>
 
   <script>
-    let edit_url = "{% url 'projects-id-metadata-id-update' table_id=0 %}"
+    let edit_url = "{% url 'projects-id-metadata-create'%}?table_id=0&step=3"
     let upload_url = "{% url 'projects-id-metadata-id-data-update' table_id=0 %}"
 
     function rowSelected(row) {
